% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_glmer.R
\name{cc_glmer}
\alias{cc_glmer}
\title{cc_glmer - Funktion zur Erstellung von Modellen und Vergleich von Modellen}
\usage{
cc_glmer(target, effects, data, test = "Chisq")
}
\arguments{
\item{target}{Zielvariable}

\item{effects}{Liste der Effekte}

\item{data}{Datenrahmen}

\item{test}{Teststatistik}
}
\value{
Eine Liste mit ANOVA-Tabelle, Modellliste und Zusammenfassungsliste
}
\description{
Diese Funktion erstellt Modelle und vergleicht sie für eine gegebene Zielvariable und eine Liste von Effekten.
}
\seealso{
\code{\link[lme4]{glmer}}, \code{\link[stats]{anova}}
}
