% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_lm.R
\name{cc_lm}
\alias{cc_lm}
\title{cc_lm - Funktion zur Vergleichsanalyse von linearen Modellen}
\usage{
cc_lm(target, effects, data, test = "Chisq")
}
\arguments{
\item{target}{Zielvariable}

\item{effects}{Liste der Effekte}

\item{data}{Datenrahmen}

\item{test}{Teststatistik}
}
\value{
Eine Liste, die ANOVA-Tabelle, die model_list und die summary_list enthält
}
\description{
Diese Funktion führt eine Vergleichsanalyse von linearen Modellen für eine gegebene Zielvariable und eine Liste von Effekten durch.
}
\examples{
# Erstellen Sie eine Liste von Effekten
effects <- list("cyl", "disp")
cc_lm(target = "mpg", effects = effects, data = mtcars, test = "Chisq")
}
\seealso{
\code{\link[stats]{lm}}, \code{\link[stats]{anova}}
}
