% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_lmer.R
\name{cc_lmer}
\alias{cc_lmer}
\title{cc_lmer - Funktion zur Erstellung von linearen gemischten Effektmodellen}
\usage{
cc_lmer(target, effects, data, test = "Chisq")
}
\arguments{
\item{target}{Zielvariable als Zeichenfolge}

\item{effects}{Liste der Effekte}

\item{data}{Datenrahmen}

\item{test}{Teststatistik}
}
\value{
Eine Liste, die ANOVA-Tabelle, die Modellliste und die Zusammenfassungsliste enthält
}
\description{
Diese Funktion erstellt lineare gemischte Effektmodelle für eine gegebene Zielvariable und eine Liste von Effekten.
}
\seealso{
\code{\link[lme4]{lmer}}, \code{\link[stats]{anova}}
}
