% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_assumptions.R
\name{check_model_assumptions}
\alias{check_model_assumptions}
\title{Überprüft die Modellannahmen}
\usage{
check_model_assumptions(model)
}
\arguments{
\item{model}{Ein Modellobjekt, das überprüft werden soll.}
}
\value{
Eine Liste mit den Ergebnissen der Normalitätstests für die Residuen, dem Breusch-Pagan-Test für die Heteroskedastizität der Residuen und dem Variationsinflationfaktor für die Multikollinearität der Prädiktoren.
}
\description{
Diese Funktion überprüft die Modellannahmen für das gegebene Modell und gibt eine Liste mit den Ergebnissen zurück.
}
