% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_boxplots.R
\name{des_boxplots}
\alias{des_boxplots}
\title{Create boxplots for numeric columns in a data frame}
\usage{
des_boxplots(
  data,
  x_var = "origin",
  color_var = NULL,
  title_prefix = "Boxplot of",
  theme_options = list(),
  facet_var = NULL,
  color_values = NULL,
  save_plots = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{x_var}{A character string specifying the variable to be plotted on the x-axis. Default is "origin".}

\item{color_var}{A character string specifying the variable to be used for coloring the boxplots. Default is NULL.}

\item{title_prefix}{A character string specifying the prefix for the plot titles. Default is "Boxplot of".}

\item{theme_options}{A named list of ggplot2 theme options to be applied to the plots. Default is an empty list.}

\item{facet_var}{A character string specifying the variable to be used for faceting the boxplots. Default is NULL.}

\item{color_values}{A vector of color values to be used for coloring the boxplots. Default is NULL.}

\item{save_plots}{A logical value indicating whether the plots should be saved as .jpg files. Default is FALSE.}
}
\value{
A list of ggplot2 objects containing the boxplots.
}
\description{
This function creates boxplots for all numeric columns in a data frame. The boxplots can be colored by a specified variable and faceted by another variable.
}
