% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_histo.R
\name{des_hist}
\alias{des_hist}
\title{Create custom histograms based on user inputs}
\usage{
des_hist(
  input,
  type = "specific",
  hist_column = NULL,
  value_column = NULL,
  id_name = c(method_id = "hist_id"),
  save_plots = FALSE,
  to_csv = FALSE,
  output_file = "data.csv",
  color_var = NULL,
  color_values = NULL,
  facet_var = NULL
)
}
\arguments{
\item{input}{A data frame containing the data to plot.}

\item{type}{A character string indicating the type of histograms to create.
Possible values are "all", "specific", or "id_only". Default is "specific".}

\item{hist_column}{A character string giving the name of the column in the input
data frame that contains the values to plot. Required if type is "specific" or
"id_only".}

\item{value_column}{A character string giving the name of the column in the input
data frame that contains the values to plot. Required if type is "all".}

\item{id_name}{A named character vector giving the new name for the column that
contains the histogram identifier. Default is c(method_id = "hist_id").}

\item{save_plots}{A logical value indicating whether to save the histograms as
.jpg files in a folder named "histograms". Default is FALSE.}

\item{to_csv}{A logical value indicating whether to save the processed data frame
to a CSV file. Default is FALSE.}

\item{output_file}{A character string giving the name of the CSV file to save the
data to. Default is "data.csv".}

\item{color_var}{A character string giving the name of the column in the input
data frame that contains the variable to use for coloring the histograms.
Optional.}

\item{color_values}{A named character vector giving the custom color values to use
for the color_var. Optional.}

\item{facet_var}{A character string giving the name of the column in the input
data frame that contains the variable to use for faceting the histograms.
Optional.}
}
\value{
A list containing the processed data frame and the histograms.
}
\description{
This function creates histograms for one or more numeric columns in a data frame,
using the ggplot2 package. The user can specify the column(s) to plot, the color
variable, the facet variable, and other options. The function also creates a new
column with a unique identifier for each histogram, which can be used for further
analysis or saved to a CSV file.
}
\examples{
\dontrun{
des_hist(hessen_means_select, type = "all",  hist_column = "stoff",
         value_column = "ergebnis", id_name = c(method_id = "hist_id"),
         save_plots = FALSE, to_csv = FALSE, output_file = "data.csv",
         color_var = "methode", color_values = NULL, facet_var = NULL)
}
}
