% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_shapiro.R
\name{des_shapiro}
\alias{des_shapiro}
\title{Perform Shapiro-Wilk test on all numeric variables in a data frame}
\usage{
des_shapiro(
  input,
  to_csv = FALSE,
  drop_cols = NULL,
  output_file = "output.csv"
)
}
\arguments{
\item{input}{A data frame or a tibble.}

\item{to_csv}{A logical value indicating whether to save the results to a CSV
file or not. Default is FALSE.}

\item{drop_cols}{A character vector of column names to be dropped from the
data frame before testing. Default is NULL.}

\item{output_file}{A character string specifying the name of the output file
if to_csv is TRUE. Default is "output.csv".}
}
\value{
A data frame with three columns: data, W, and p_value. The data
  column contains the names of the numeric variables in the input data frame.
  The W column contains the Shapiro-Wilk test statistic for each variable.
  The p_value column contains the p-value for each variable.
}
\description{
This function performs the Shapiro-Wilk test of normality on all numeric
variables in a given data frame. It returns a data frame with the test
statistic and p-value for each variable. It also allows the user to save the
results to a CSV file, and to drop some columns from the data frame before
testing.
}
\examples{
# Use the mtcars data set as an example
data(mtcars)

# Run the des_shapiro function on the mtcars data set
des_shapiro(mtcars)

# Save the results to a file named "mtcars_shapiro.csv"
des_shapiro(mtcars, to_csv = TRUE, output_file = "mtcars_shapiro.csv")

# Drop the vs and am columns before testing
des_shapiro(mtcars, drop_cols = c("vs", "am"))

}
