% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_data2.R
\name{read_and_clean_data2}
\alias{read_and_clean_data2}
\title{Read and clean data from a CSV file}
\usage{
read_and_clean_data2(file_path, seperate = ";", decim = ",", tibble = TRUE)
}
\arguments{
\item{file_path}{A character string specifying the path to the CSV file.}

\item{seperate}{A character string specifying the separator used in the CSV file. Default is ";".}

\item{decim}{A character string specifying the decimal point used in the CSV file. Default is ",".}

\item{tibble}{A logical value indicating whether the output should be a tibble. Default is TRUE.}
}
\value{
A tibble containing the cleaned data.
}
\description{
This function reads a CSV file with specified separator and decimal point, cleans the column names using the clean_names() function from the janitor package, and returns the cleaned data as a tibble.
It also removes empty rows. See ??read_and_clean_data() as the previous version where more cleaning steps were performed.
}
