% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_changer.R
\name{row_changer}
\alias{row_changer}
\title{Rename Rows in Data Frame}
\usage{
row_changer(
  input,
  type = NULL,
  origin_col = NULL,
  new_col = NULL,
  from = NULL,
  to = NULL,
  overwrite = FALSE,
  tibble = TRUE,
  output_file = "row_rename/renamed_input.csv",
  to_csv = FALSE
)
}
\arguments{
\item{input}{The input data frame.}

\item{type}{The type of row renaming. Must be set to either 'all' or 'specific'.
This determines where to rename rows. Check help(row_rename) for additional explanations.}

\item{origin_col}{The name of the column that contains the original values.
This argument is required when type is set to 'specific'.}

\item{new_col}{The name of the column where the renamed values should be stored.
This argument is required when type is set to 'specific' and overwrite is set to FALSE.}

\item{from}{A character vector of patterns to be replaced.}

\item{to}{A character vector of replacement strings.}

\item{overwrite}{Should the original column be overwritten with the renamed values?
This argument is only applicable when type is set to 'specific' and new_col is not provided.
If TRUE, the original column will be overwritten with the renamed values. If FALSE, a new column
will be created with the renamed values.}

\item{tibble}{Should the input data frame be converted to a tibble?
Defaults to TRUE.}

\item{output_file}{The file path where the renamed data frame should be saved as a CSV file.
Defaults to "row_rename/renamed_input.csv".}

\item{to_csv}{Should the renamed data frame be saved as a CSV file?
Defaults to FALSE.}
}
\description{
Rename rows in a data frame based on specified patterns and replacements.
}
