% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tra_env.R
\name{tra_env}
\alias{tra_env}
\title{Import environmental data
This function imports environmental data and creates new columns containing the name of the state and the env_id which is a paste object of the state, the bdf, the year and the month.}
\usage{
tra_env(
  input,
  output_file = glue::glue("import/env_import_{state_name}.csv"),
  str_output = "str.csv",
  state_name = "Bundesland",
  to_csv = FALSE
)
}
\arguments{
\item{input}{A tibble containing the environmental data.}

\item{output_file}{The path to the output file. Default is "import/env_import_{state_name}.csv".}

\item{str_output}{The path to the output file for the structured data frame. Default is “str.csv".}

\item{state_name}{The name of the state. Default is "Bundesland".}

\item{to_csv}{A logical value indicating whether to create a csv file. Default is FALSE.}
}
\value{
A list containing two objects: env_import and str_env_import.
}
\description{
Import environmental data
This function imports environmental data and creates new columns containing the name of the state and the env_id which is a paste object of the state, the bdf, the year and the month.
}
