% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tra_meta.R
\name{tra_meta}
\alias{tra_meta}
\title{Transform data and generate metadata}
\usage{
tra_meta(
  input,
  cols_group = NULL,
  index_output = "metadata/group_indices_df.csv",
  vars_output = "metadata/group_vars_df",
  firstrows_output = "metadata/first_rows_df.csv",
  to_csv = FALSE
)
}
\arguments{
\item{input}{A data frame or a tibble to be transformed and analyzed.}

\item{cols_group}{A character vector of column names that shall be grouped by.}

\item{index_output}{A character string of the file path where the group indices data frame will be written to, if to_csv is TRUE. Default is "metadata/group_indices_df.csv".}

\item{vars_output}{A character string of the file path where the group variables data frame will be written to, if to_csv is TRUE. Default is "metadata/group_vars_df".}

\item{firstrows_output}{A character string of the file path where the first rows data frame will be written to, if to_csv is TRUE. Default is "metadata/first_rows_df.csv".}

\item{to_csv}{A logical value indicating whether to write the data frames to CSV files or not. Default is FALSE.}
}
\value{
A list of three data frames: group_indices_df, group_vars_df, and first_rows_df.
}
\description{
This function takes a data frame or a tibble as input and performs some transformations and metadata generation based on the specified columns to group by. It returns a list of three data frames: one with the group indices, one with the group variables, and one with the first rows of each group. Optionally, it can also write the data frames to CSV files in a specified directory.
It can be used to describe the grouped data frames when you are presenting a grouped transformation.
}
