% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tra_process.R
\name{tra_process}
\alias{tra_process}
\title{Process data and calculate means}
\usage{
tra_process(
  input_data,
  output_file = "meandata/means.csv",
  cols_drop = NULL,
  cols_rename = NULL,
  cols_group = NULL,
  to_csv = FALSE
)
}
\arguments{
\item{input_data}{A data frame to be processed.}

\item{output_file}{A character string specifying the path and name of the output CSV file. Default is "meandata/means.csv".}

\item{cols_drop}{A character vector of column names to be dropped from the input data. Default is NULL.}

\item{cols_rename}{A named character vector of new names for the columns to be renamed. For example, c(old_name = "new_name"). Default is NULL.}

\item{cols_group}{A character vector of column names to be used for grouping the data. Default is NULL.}

\item{to_csv}{A logical value indicating whether to write the processed data to a CSV file. Default is FALSE.}
}
\value{
A tibble of the processed data with means of numeric columns and first values of character columns.
}
\description{
This function converts the data frame to a tibble, optionally drops and renames columns, groups the data by specified columns and calculates the mean of numeric columns, and optionally writes the processed data to a CSV file.
}
\examples{
# Load the iris data set
data(iris)

# Process the iris data with tra_process function
tra_process(iris, cols_drop = "Species", cols_rename = c(Sepal.Length = "sepal_length"), cols_group = c("Petal.Width", "Petal.Length"), to_csv = FALSE)
}
