% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tra_select_means.R
\name{tra_select_means}
\alias{tra_select_means}
\title{Selects and formats data from a tibble}
\usage{
tra_select_means(
  input,
  output_file = "means_bundesland/means_bundesland.csv",
  selects = NULL,
  date_format = c("dmy", "mdy", "ymd", "dBy", "mdBy", "ymdHMS"),
  date_col = "datum",
  to_csv = FALSE
)
}
\arguments{
\item{input}{A tibble containing data to be selected and formatted.}

\item{output_file}{A character string specifying the name of the output file. Default is "means_bundesland/means_bundesland.csv".}

\item{selects}{A character vector specifying the names of the columns to be selected. Default is NULL.}

\item{date_format}{A character vector specifying the format of the date column. Default is c("dmy", "mdy", "ymd", "dBy", "mdBy", "ymdHMS").}

\item{date_col}{A character string specifying the name of the date column. Default is "datum".}

\item{to_csv}{A logical value indicating whether to write the output to a CSV file. Default is FALSE.}
}
\value{
A list containing the selected and formatted data and its structure.
}
\description{
This function selects columns from a tibble and formats the data by adding year and month columns.
}
\examples{
tra_select_means(input = data, selects = c("col1", "col2"), date_col = "date", to_csv = TRUE)

}
