% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tra_switch.R
\name{tra_switch}
\alias{tra_switch}
\title{Switch}
\usage{
tra_switch(
  data,
  type = NULL,
  variable.name = "time",
  value.var = "value",
  melts_long = "x\\\\d+",
  set_numeric = TRUE,
  ids = NULL,
  time_pattern = "x\\\\d+",
  time_prefix = "x",
  time = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{type}{A character string. 'long' means it will be switched from many columns to two new columns, one contains the values and one contains the original column names, this is only necessary if the columns that shall be melted contain any pattern in their name, e.g. starts with x & ends with a number; 'long_melt' means that the names of the columns that shall be melted do not contain any pattern, they can be named totally different, here it is necessary to set the id columns (see ids); 'short' is the opposite of long, so it 'demelts' the data, sometimes this is useful e.g. for linear models;}

\item{variable.name}{A character string. It is the name of the new column that contains the column names of the melted variables, default is set to 'time'.}

\item{value.var}{A character string. It is the name of the new column that contains the values of the melted variables, default is set to 'value'.}

\item{melts_long}{A character string. It is the pattern of the column names that shall be melted (see type = long), default is set to 'x\\d+'.}

\item{set_numeric}{If TRUE time column will be set into numeric, which is often necessary for creating plots}

\item{ids}{A numeric vector. It is the id's to colindizes when 'type = long_melt', determines which columns will stay (see type), default is set to NULL.}

\item{time_pattern}{A character string. It is the pattern of the time information in the column names that match the melts_long pattern, default is set to 'x\\d+'.}

\item{time_prefix}{A character string. It is the prefix of the time information in the column names that match the melts_long pattern, default is set to 'x'.}

\item{time}{A numeric vector. It describes the vector of the time values that correspond to the time information in the column names that match the melts_long pattern, default is set to NULL.}
}
\description{
Switch
}
\examples{
data(mtcars)
result <- tra_switch(data = mtcars, type = "long_melt", variable.name = "variable", value.var = "value", ids = 1:3)
}
